requirements.txt
wheel
python-dotenv
requests
flask